herbal_article_creator/requirements.txt
#Agentic AI Engine
crewai[tools]>=0.201.1

#Search
tavily-python>=0.5.0

# LangChain
langchain>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0

# Vector Store
chromadb>=0.4.22
pinecone>=4.0.0

# Google AI
google-generativeai>=0.3.0

# PDF Processing (if using PDFs)
pymupdf>=1.23.0
pypdfium2>=4.26.0
pdfplumber>=0.10.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Optional but recommended
requests>=2.31.0

# === Core Google APIs ===
google-api-python-client>=2.136.0
google-auth>=2.35.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# === PDF support ===
pdfplumber>=0.11.4
PyPDF2>=3.0.1

# === Optional OCR (ถ้าอยากรองรับ PDF scan) ===
pytesseract>=0.3.10
Pillow>=10.0.0

# for markdown → HTML + convert to .docx
python-docx>=1.1.0
markdown>=3.5.2
beautifulsoup4>=4.12.3
lxml>=5.3.0

# Google Drive integration
pydrive2>=1.19.0

#Zep memory integration
python-slugify==8.0.4
text-unidecode==1.3
zep-cloud==3.5.0
zep-crewai==1.1.1

#check pip
#pip list | grep langchain

#Evaluation
#pip install rouge-score sacrebleu
# absl-py==2.3.1
# colorama==0.4.6
# joblib==1.5.2
# nltk==3.9.2
# rouge-score==0.1.2
# sacrebleu==2.5.1
# tabulate==0.9.0

rouge-score==0.1.2
sacrebleu==2.4.2

#performance crewAI
psutil==7.1.2

# uv pip install selenium webdriver-manager
#  + outcome==1.3.0.post0
#  + pysocks==1.7.1
#  + selenium==4.38.0
#  + sortedcontainers==2.4.0
#  + trio==0.32.0
#  + trio-websocket==0.12.2
#  - urllib3==2.3.0
#  + urllib3==2.5.0
#  + webdriver-manager==4.0.2
#  + wsproto==1.2.0

# uv pip install biopython

# uv pip install langtrace-python-sdk
#  + boto3==1.38.0
#  + botocore==1.38.46
#  + deprecated==1.2.18
#  + jmespath==1.0.1
#  + langtrace-python-sdk==3.8.21
#  - opentelemetry-api==1.37.0
#  + opentelemetry-api==1.38.0
#  - opentelemetry-exporter-otlp-proto-common==1.37.0
#  + opentelemetry-exporter-otlp-proto-common==1.38.0
#  - opentelemetry-exporter-otlp-proto-grpc==1.37.0
#  + opentelemetry-exporter-otlp-proto-grpc==1.38.0
#  - opentelemetry-exporter-otlp-proto-http==1.37.0
#  + opentelemetry-exporter-otlp-proto-http==1.38.0
#  + opentelemetry-instrumentation==0.59b0
#  + opentelemetry-instrumentation-sqlalchemy==0.59b0
#  - opentelemetry-proto==1.37.0
#  + opentelemetry-proto==1.38.0
#  - opentelemetry-sdk==1.37.0
#  + opentelemetry-sdk==1.38.0
#  - opentelemetry-semantic-conventions==0.58b0
#  + opentelemetry-semantic-conventions==0.59b0
#  + s3transfer==0.12.0
#  + safetensors==0.6.2
#  + sentry-sdk==2.45.0
#  + setuptools==80.9.0
#  + trace-attributes==7.2.1
#  + transformers==4.57.1
#  + ujson==5.11.0
#  + wrapt==1.17.3

#uv pip install langfuse
#+ langfuse==3.10.0